Test cases can no longer be abstract
